Added Logout Buttons
