+++ apps/web/src/components/AccessDenied.tsx
@@ -1,8 +1,19 @@
 import React from 'react';
-import { Box, Typography, Container } from '@mui/material';
-import { Block as BlockIcon } from '@mui/icons-material';
+import { Box, Typography, Container, Button } from '@mui/material';
+import { Block as BlockIcon, Logout as LogoutIcon } from '@mui/icons-material';
+import { useAuth } from '../contexts/AuthContext';
 
 const AccessDenied: React.FC = () => {
+  const { logout } = useAuth();
+
+  // Handle logout functionality
+  const handleLogout = () => {
+    logout();
+    // Redirect to login page with proper redirect_uri
+    const redirectUri = encodeURIComponent('http://localhost:3001/payments');
+    window.location.href = `https://grupo5-usuarios.vercel.app/login?redirect_uri=${redirectUri}`;
+  };
+
   return (
     <Container maxWidth="sm">
       <Box
@@ -39,16 +50,34 @@ const AccessDenied: React.FC = () => {
           color="text.secondary"
           sx={{ mb: 3, maxWidth: '400px' }}
         >
-          No tienes permisos suficientes para acceder a esta funcionalidad.
+          No tenés permisos suficientes para acceder a esta funcionalidad.
         </Typography>
         
         <Typography 
           variant="body2" 
           color="text.disabled"
-          sx={{ fontStyle: 'italic' }}
+          sx={{ fontStyle: 'italic', mb: 4 }}
         >
           Contacta al administrador si crees que esto es un error.
         </Typography>
+
+        <Button
+          variant="contained"
+          size="large"
+          startIcon={<LogoutIcon />}
+          onClick={handleLogout}
+          sx={{ 
+            textTransform: 'none',
+            fontWeight: 600,
+            minWidth: '200px',
+            backgroundColor: 'error.main',
+            '&:hover': {
+              backgroundColor: 'error.dark'
+            }
+          }}
+        >
+          Cerrar Sesión
+        </Button>
       </Box>
     </Container>
   );
